codie/auth_service/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
pyjwt
bcrypt 